Change tag for representatives api endpoints to Representative
