+++ api/client.yaml
@@ -1532,7 +1532,7 @@ paths:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
   /customers/{customerID}/representatives:
     post:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Add Customer Representative
       description: Add a Customer Representative
       operationId: addRepresentative
@@ -1578,7 +1578,7 @@ paths:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
   /customers/{customerID}/representatives/{representativeID}:
     put:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Update Customer Representative
       description: Updates the specified Customer Representative
       operationId: updateRepresentative
@@ -1613,7 +1613,7 @@ paths:
               schema:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
     delete:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Delete Customer Representative
       description: Deletes a Customer Representative
       operationId: deleteRepresentative
@@ -1643,7 +1643,7 @@ paths:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
   /customers/{customerID}/representatives/{representativeID}/address:
     post:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Add Customer Representative Address
       description: Add an address to an existing customer representative record
       operationId: addRepresentativeAddress
@@ -1695,7 +1695,7 @@ paths:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
   /customers/{customerID}/representatives/{representativeID}/addresses/{addressID}:
     put:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Update Customer Representative Address
       description: Updates the specified customer representative address
       operationId: updateRepresentativeAddress
@@ -1737,7 +1737,7 @@ paths:
               schema:
                 $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
     delete:
-      tags: [Customers]
+      tags: [Representatives]
       summary: Delete a Customer Representative Address
       description: Deletes a customer representative's address
       operationId: deleteRepresentativeAddress
