--- pkg/client/docs/CustomerAddress.md
@@ -1,19 +1,0 @@
-# CustomerAddress
-
-## Properties
-
-Name | Type | Description | Notes
------------- | ------------- | ------------- | -------------
-**AddressID** | **string** | Unique identifier for this Address | 
-**Type** | [**AddressType**](AddressType.md) |  | 
-**Address1** | **string** | First line of the address | 
-**Address2** | **string** | Second line of the address | [optional] 
-**City** | **string** |  | 
-**State** | **string** | two character code of a US state | 
-**PostalCode** | **string** |  | 
-**Country** | **string** |  | 
-**Validated** | **bool** | Address has been validated for customer | [optional] 
-
-[[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
-
-
